Extract class for Practice Controller
